+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `/filesystem/cluster-slurm-ssh/ops/view` endpoint now accepts `size` and `offset` parameters to arbitrary select a chunk of a file
+
 ### Fixed
 
 ## [2.2.8]